+++ com.unity.shadergraph/Editor/Generation/Targets/BuiltIn/Editor/AssetPostProcessors/ShaderGraphMaterialsUpdater.cs
@@ -1,13 +1,5 @@
-using UnityEditor.Rendering.BuiltIn;
 using UnityEditor.ShaderGraph;
 using UnityEngine;
-
-// The purpose of this file is to handle when a BuiltIn ShaderGraph is saved,
-// and ensure all Materials that use that ShaderGraph
-// have updated their keywords and other settings appropriately.
-// This system should be removed once we can declare a Material to have
-// an import dependency on the ShaderGraph artifact
-// (which would have the same effect, but work faster)
 
 namespace UnityEditor.Rendering.BuiltIn
 {
@@ -19,8 +11,6 @@ namespace UnityEditor.Rendering.BuiltIn
     [InitializeOnLoad]
     class ShaderGraphMaterialsUpdater
     {
-        const string kMaterialFilter = "t:Material";
-
         static ShaderGraphMaterialsUpdater()
         {
             GraphData.onSaveGraph += OnShaderGraphSaved;
@@ -35,11 +25,11 @@ namespace UnityEditor.Rendering.BuiltIn
             if (!builtInShaderGraphSaveContext.updateMaterials)
                 return;
 
-            // Iterate all Materials
-            string[] materialGuids = AssetDatabase.FindAssets(kMaterialFilter);
+            // Iterate over all loaded Materials
+            Material[] materials = Resources.FindObjectsOfTypeAll<Material>();
             try
             {
-                for (int i = 0, length = materialGuids.Length; i < length; i++)
+                for (int i = 0, length = materials.Length; i < length; i++)
                 {
                     // Only update progress bar every 10 materials
                     if (i % 10 == 9)
@@ -50,25 +40,14 @@ namespace UnityEditor.Rendering.BuiltIn
                             i / (float)(length - 1));
                     }
 
-                    // Get Material object
-                    string materialPath = AssetDatabase.GUIDToAssetPath(materialGuids[i]);
-                    Material material = AssetDatabase.LoadAssetAtPath<Material>(materialPath);
-
                     // Reset keywords
-                    if ((material != null) && material.shader.name == shader.name)
-                        ShaderUtils.ResetMaterialKeywords(material);
-
-                    material = null;
-
-                    // Free the materials every 200 iterations, on big project loading all materials in memory can lead to a crash
-                    if ((i % 200 == 0) && i != 0)
-                        EditorUtility.UnloadUnusedAssetsImmediate(true);
+                    if (materials[i].shader.name == shader.name)
+                        ShaderUtils.ResetMaterialKeywords(materials[i]);
                 }
             }
             finally
             {
                 EditorUtility.ClearProgressBar();
-                EditorUtility.UnloadUnusedAssetsImmediate(true);
             }
         }
     }